2913: Refactor and cleanup
